+++ packages/document-schema/type-gen/src/lib/pack-docschema-api/pack-docschema-ir/schemaMeta.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { ISchemaVersion } from "../pack-docschema-api/schemaVersion";
+import type { ISchemaVersion } from "../pack-docschema-api/schemaVersion.js";
 export interface ISchemaMeta {
   readonly "addedIn": ISchemaVersion;
   readonly "deprecated"?: ISchemaVersion | null;