+++ packages/monorepo/transpile/CHANGELOG.md
@@ -1,3 +1,7 @@
 # @palantir/pack.monorepo.transpile
 
+## 0.0.4
+
+## 0.0.3
+
 ## 0.0.2