+++ packages/monorepo/tsconfig/CHANGELOG.md
@@ -1,3 +1,7 @@
 # @palantir/pack.monorepo.tsconfig
 
+## 0.4.2
+
+## 0.4.1
+
 ## 0.4.0